requirements.txt
langchain
ipykernel
python-dotenv
langchain_community
pypdf
bs4
arxiv
pymupdf
Wikipedia
langchain_text_splitters
#langchain_ollama
langchain-huggingface
faiss-cpu
langchain_chroma
selenium 
webdriver-manager
langchain_ollama
langchain_openai
streamlit
langchain-groq
fastapi
uvicorn
langserve
# sse_starlette
# atlassian-python-api
#lxml
#pytesseract
#pillow
# tesseract
# pdf2image
# python-poppler
#mysql-connector-python
#pymysql
#pyodbc
langchain_experimental

#Okta
authlib
pkce
fastapi

scikit-learn
# pdf2image
# pytesseract
pdfplumber

langchain-groq